+++ src/client/cypress/e2e/editor/instrumentation.cy.js
@@ -5,6 +5,20 @@ import {
   login,
 } from "../testHelpers";
 
+const delayedType = (selector, string) => {
+  const element = cy.get(selector);
+  cy.wait(500);
+  element.type(string);
+};
+
+const addInstrumentLayer = () => {
+  cy.wait(1000);
+  cy.get('[data-cy="add-instrumentation-button"]').click({ force: true });
+  cy.wait("@stratigraphy_layer_edit_create");
+  cy.wait("@layer");
+  cy.wait(5000);
+};
+
 describe("Instrumentation tests", () => {
   beforeEach(() => {
     interceptApiCalls();
@@ -17,7 +31,7 @@ describe("Instrumentation tests", () => {
   });
 
   afterEach(() => {
-    // delete borehole
+    // Delete borehole
     cy.get("@borehole_id").then(id => deleteBorehole(id));
   });
 
@@ -32,21 +46,20 @@ describe("Instrumentation tests", () => {
       "contain",
       "For the recording of an instrument please click the plus symbol at the top left",
     );
+    cy.contains("a", "Stop editing").click();
   });
 
   it("Can add and delete instrument without completion", () => {
     cy.contains("a", "Start editing").click();
     cy.wait("@edit_lock");
 
-    // header should not contain tabs
+    // Header should not contain tabs
     cy.get('[data-cy="profile-header-list"]')
       .children()
       .should("have.length", 0);
 
-    // add instrument with completion "no casing"
-    cy.get('[data-cy="add-instrumentation-button"]').click();
-    cy.wait("@edit_list");
-    cy.wait("@layer");
+    addInstrumentLayer();
+    cy.wait(500);
 
     cy.get('[data-cy="casingName"]')
       .children()
@@ -57,15 +70,16 @@ describe("Instrumentation tests", () => {
         cy.wrap(el).click().find('[role="option"]').last().click(),
       );
 
-    // header should contain one tab with name "No casind"
+    // Header should contain one tab with name "No casing"
     cy.get('[data-cy="profile-header-list"]')
       .children()
       .should("have.length", 1);
 
-    // add another instrument without name of completion
-    cy.get('[data-cy="add-instrumentation-button"]').click();
+    // Add another instrument without name of completion
+    addInstrumentLayer();
+    cy.wait(500);
 
-    // select tab "No completion"
+    // Select tab "No completion"
     cy.get('[data-cy="profile-header-list"]')
       .children()
       .first()
@@ -74,19 +88,168 @@ describe("Instrumentation tests", () => {
       .should("contain", "No casing")
       .click();
 
-    // only one instrument should be displayed
+    // Only one instrument should be displayed
     cy.get('[data-cy="instrument-list"]').children().should("have.length", 1);
 
-    // show all instruments
+    // Show all instruments
     cy.get('[data-cy="show-all-button"]').click();
     cy.get('[data-cy="instrument-list"]').children().should("have.length", 2);
 
-    // delete instrument with "no completion" removes "No completion" tab.
+    // Delete instrument with "no completion" removes "No completion" tab.
     cy.get('[data-cy="delete-instrument-button"]').first().click();
 
-    // header should not contain tabs
+    // Header should not contain tabs
     cy.get('[data-cy="profile-header-list"]')
       .children()
       .should("have.length", 0);
+
+    cy.contains("a", "Stop editing").click();
+  });
+
+  it("Can add casing layer to instrument", () => {
+    cy.contains("a", "Start editing").click();
+    cy.wait("@edit_lock");
+
+    // Add two casings with two layers
+    cy.get('[data-cy="casing-menu-item"]').click();
+
+    // First casing
+    cy.get('[data-cy="add-stratigraphy-button"]').click();
+    cy.wait("@stratigraphy_edit_create");
+    cy.wait("@layer");
+    cy.contains("div", "This is the main completion");
+
+    delayedType('[data-cy="name"]', "Moonshine Bike");
+
+    cy.get('[data-cy="add-layer-button"]').click();
+    cy.wait("@layer");
+    cy.get('[data-cy="styled-layer-0"]').click();
+    delayedType('[data-cy="casing_id"]', "Moonshine Veal");
+
+    cy.get('[data-cy="add-layer-button"]').click();
+    cy.wait("@layer");
+    cy.get('[data-cy="styled-layer-1"]').click();
+    delayedType('[data-cy="casing_id"]', "Moonshine Honey");
+
+    // Second casing
+    cy.get('[data-cy="add-stratigraphy-button"]').click();
+    cy.wait("@stratigraphy_edit_create");
+    cy.wait("@layer");
+
+    // Navigate to correct tab
+    cy.contains("div", "Unknown").click();
+    delayedType('[data-cy="name"]', "Sunshine Bike");
+
+    cy.get('[data-cy="add-layer-button"]').click();
+    cy.wait("@layer");
+    cy.wait("@stratigraphy_layer_edit_create");
+
+    cy.get('[data-cy="styled-layer-0"]').click();
+    delayedType('[data-cy="casing_id"]', "Sunshine Veal");
+
+    cy.get('[data-cy="add-layer-button"]').click();
+    cy.wait("@layer");
+    cy.wait("@stratigraphy_layer_edit_create");
+
+    cy.get('[data-cy="styled-layer-1"]').scrollIntoView().click();
+    delayedType('[data-cy="casing_id"]', "Sunshine Honey");
+
+    // Add instrument
+    cy.get('[data-cy="instrument-menu-item"]').click();
+    cy.wait("@layer");
+
+    addInstrumentLayer();
+
+    // Chose first casing
+    let casingDropDown = cy
+      .get('[data-cy="casingName"]')
+      .children()
+      .first()
+      .children()
+      .first();
+
+    casingDropDown.each((el, index, list) =>
+      cy
+        .wrap(el)
+        .click({ force: true })
+        .find('[role="option"]')
+        .eq(2)
+        .click({ force: true }),
+    );
+
+    casingDropDown.contains("Moonshine Bike");
+
+    cy.wait("@casing-layers");
+    cy.wait(3000);
+
+    // Choose second casingLayer
+    let casingLayerDropDown = cy
+      .get('[data-cy="casingId"]')
+      .children()
+      .first()
+      .children()
+      .first();
+
+    casingLayerDropDown.each((el, index, list) =>
+      cy
+        .wrap(el)
+        .click({ force: true })
+        .find('[role="option"]')
+        .eq(2)
+        .click({ force: true }),
+    );
+
+    casingLayerDropDown.contains("Moonshine Honey");
+
+    casingDropDown = cy
+      .get('[data-cy="casingName"]')
+      .children()
+      .first()
+      .children()
+      .first();
+
+    // Choose second casing
+    casingDropDown.each((el, index, list) =>
+      cy
+        .wrap(el)
+        .click({ force: true })
+        .find('[role="option"]')
+        .eq(3)
+        .click({ force: true }),
+    );
+
+    cy.wait("@casing-layers");
+    cy.wait(3000);
+
+    casingDropDown.contains("Sunshine Bike");
+
+    casingLayerDropDown = cy
+      .get('[data-cy="casingId"]')
+      .children()
+      .first()
+      .children()
+      .first();
+
+    // Change of casing should reset casingLayer
+    casingLayerDropDown.should("not.contain", "Moonshine Bike");
+
+    cy.wait("@casing-layers");
+    cy.wait(3000);
+
+    // Dropdown options in casingLayer dropdown have updated
+    casingLayerDropDown.each((el, index, list) =>
+      cy
+        .wrap(el)
+        .click({ force: true })
+        .find('[role="option"]')
+        .eq(2)
+        .click({ force: true }),
+    );
+
+    cy.wait("@casing-layers");
+    cy.wait(3000);
+
+    casingLayerDropDown.contains("Sunshine Honey");
+    cy.contains("a", "Stop editing").click();
   });
 });